Refactor: move generateToken to utils
