+++ backend/utils/generateToken.js
